LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/top.http-rp/same-origin/xhr.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=7ee705d3-995b-4884-a8fb-8469b515d53b&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=7ee705d3-995b-4884-a8fb-8469b515d53b&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=7d6e6722-ac90-4ec3-aa49-2aadf6ae84f7&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=7d6e6722-ac90-4ec3-aa49-2aadf6ae84f7&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=9b57424e-6287-4bcf-b3f7-255635994777&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=9b57424e-6287-4bcf-b3f7-255635994777&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=0bbf2b36-22f8-4dac-8294-19a7403378ae&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=0bbf2b36-22f8-4dac-8294-19a7403378ae&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=1cc2ee05-7953-4fbe-ba05-20c9141ff24b&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=1cc2ee05-7953-4fbe-ba05-20c9141ff24b&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=8f4a5ed4-74fc-4a55-84b4-14f1e17702d9&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=8f4a5ed4-74fc-4a55-84b4-14f1e17702d9&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?action=purge&key=4444551a-43f4-46a1-ab95-a9c839617951&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?action=purge&key=454ba65a-f712-4435-af0b-83b880d69083&path=%2Fmixed-content

FAIL Referrer Policy: Expects omitted for xhr to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to same-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
FAIL Referrer Policy: Expects omitted for xhr to same-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "Error: Network Error"
PASS Referrer Policy: Expects stripped-referrer for xhr to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects stripped-referrer for xhr to same-http origin and no-redirect redirection from http context.
